requirements.txt
langgraph==0.4.8
langsmith==0.3.45
langchain==0.3.25
langchain-anthropic==0.3.15
python-dotenv
langchain_openai==0.3.22
openai==1.86.0